timer

+++ src/minigames/shared/level.rs
@@ -7,7 +7,7 @@ pub struct CleanupLevel;
 pub struct Level {
     pub current_value: usize,
     pub target_score: usize,
-    pub target_time: Option<usize>,
+    pub target_time: Option<f32>,
     pub bullets: Option<usize>,
 }
